+++ app/layout.tsx
@@ -8,7 +8,6 @@ import { cn } from "@/lib/utils"
 import { Providers } from "@/components/providers"
 import { SiteHeader } from "@/components/site-header"
 import { TailwindIndicator } from "@/components/tailwind-indicator"
-import { ThemeProvider } from "@/components/theme-provider"
 
 export const metadata: Metadata = {
   title: {
