--- common/demo-data.ts
@@ -1,270 +1,0 @@
-import { ConsumptionDatum } from "./types"
-
-export const demoData: ConsumptionDatum[] = [
-  {
-    block_number: 5466551,
-    timestamp: 1706961024516,
-    ref_time: {
-      normal: 0.001,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5466552,
-    timestamp: 1706961036551,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473225,
-    timestamp: 1707041820458,
-    ref_time: {
-      normal: 0.002,
-      operational: 0,
-      mandatory: 0.021,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.005,
-    },
-  },
-  {
-    block_number: 5473226,
-    timestamp: 1707041832474,
-    ref_time: {
-      normal: 0.011,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0.001,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473227,
-    timestamp: 1707041844539,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473228,
-    timestamp: 1707041856550,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473229,
-    timestamp: 1707041868606,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473230,
-    timestamp: 1707041880843,
-    ref_time: {
-      normal: 0.001,
-      operational: 0,
-      mandatory: 0.021,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.005,
-    },
-  },
-  {
-    block_number: 5473231,
-    timestamp: 1707041892458,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473232,
-    timestamp: 1707041904643,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473233,
-    timestamp: 1707041916491,
-    ref_time: {
-      normal: 0.001,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473234,
-    timestamp: 1707041928399,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473235,
-    timestamp: 1707041940214,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.021,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.005,
-    },
-  },
-  {
-    block_number: 5473236,
-    timestamp: 1707041952868,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473237,
-    timestamp: 1707041964453,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473238,
-    timestamp: 1707041976473,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473239,
-    timestamp: 1707041988233,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-  {
-    block_number: 5473240,
-    timestamp: 1707042000248,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.021,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.005,
-    },
-  },
-  {
-    block_number: 5473241,
-    timestamp: 1707042012464,
-    ref_time: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.018,
-    },
-    proof_size: {
-      normal: 0,
-      operational: 0,
-      mandatory: 0.003,
-    },
-  },
-]
