+++ components/connect-button.tsx
@@ -3,31 +3,16 @@
 import { FC, useMemo, useState } from "react"
 import Image from "next/image"
 import Link from "next/link"
-// import { SupportedChainId } from "@azns/resolver-core"
-// import { useResolveAddressToDomain } from "@azns/resolver-react"
 import { InjectedAccount } from "@polkadot/extension-inject/types"
 import { encodeAddress } from "@polkadot/util-crypto"
 import {
-  SubstrateChain,
   SubstrateWalletPlatform,
   allSubstrateWallets,
-  getSubstrateChain,
   isWalletInstalled,
   useBalance,
   useInkathon,
 } from "@scio-labs/use-inkathon"
-import {
-  AlertOctagon,
-  ArrowDown,
-  ArrowUpRight,
-  ChevronDown,
-  LinkIcon,
-} from "lucide-react"
-import aznsIconSvg from "public/icons/azns-icon.svg"
-
-// import { AiOutlineCheckCircle, AiOutlineDisconnect } from "react-icons/ai"
-// import { FiChevronDown, FiExternalLink } from "react-icons/fi"
-// import { RiArrowDownSLine } from "react-icons/ri"
+import { ArrowUpRight, CheckCircle, ChevronDown } from "lucide-react"
 
 import { truncateHash } from "@/lib/utils"
 import { Button } from "@/components/ui/button"
@@ -38,11 +23,6 @@ import {
   DropdownMenuSeparator,
   DropdownMenuTrigger,
 } from "@/components/ui/dropdown-menu"
-import {
-  Tooltip,
-  TooltipContent,
-  TooltipTrigger,
-} from "@/components/ui/tooltip"
 
 export interface ConnectButtonProps {
   size?: "default" | "sm" | "lg" | "icon" | null | undefined
@@ -51,7 +31,6 @@ export interface ConnectButtonProps {
 export const ConnectButton: FC<ConnectButtonProps> = ({ size }) => {
   const {
     activeChain,
-    switchActiveChain,
     connect,
     disconnect,
     isConnecting,
@@ -59,20 +38,6 @@ export const ConnectButton: FC<ConnectButtonProps> = ({ size }) => {
     accounts,
     setActiveAccount,
   } = useInkathon()
-  const { reducibleBalance, reducibleBalanceFormatted } = useBalance(
-    activeAccount?.address,
-    true,
-    {
-      forceUnit: false,
-      fixedDecimals: 2,
-      removeTrailingZeros: true,
-    }
-  )
-  // const [supportedChains] = useState(
-  //   env.supportedChains.map(
-  //     (networkId) => getSubstrateChain(networkId) as SubstrateChain
-  //   )
-  // )
 
   // Sort installed wallets first
   const [browserWallets] = useState([
@@ -101,7 +66,6 @@ export const ConnectButton: FC<ConnectButtonProps> = ({ size }) => {
             size={size}
           >
             Connect Wallet
-            {/* <RiArrowDownSLine size={20} aria-hidden="true" /> */}
             <ChevronDown size={16} />
           </Button>
         </DropdownMenuTrigger>
@@ -123,7 +87,6 @@ export const ConnectButton: FC<ConnectButtonProps> = ({ size }) => {
                   <Link href={w.urls.website}>
                     <div className="align-center flex justify-start gap-2">
                       <p>{w.name}</p>
-                      {/* <FiExternalLink size={16} /> */}
                       <ArrowUpRight />
                     </div>
                     <p>Not installed</p>
@@ -138,7 +101,6 @@ export const ConnectButton: FC<ConnectButtonProps> = ({ size }) => {
   // Account Menu & Disconnect Button
   return (
     <div className="flex select-none flex-wrap items-stretch justify-center gap-4">
-      {/* Account Name, Address, and AZERO.ID-Domain (if assigned) */}
       <DropdownMenu>
         <DropdownMenuTrigger asChild className="">
           <Button className="" translate="no" size={size}>
@@ -155,11 +117,6 @@ export const ConnectButton: FC<ConnectButtonProps> = ({ size }) => {
                   )}
                 </span>
               </div>
-              {/* <FiChevronDown
-                className="shrink-0"
-                size={22}
-                aria-hidden="true"
-              /> */}
               <ChevronDown size={16} />
             </div>
           </Button>
@@ -168,28 +125,6 @@ export const ConnectButton: FC<ConnectButtonProps> = ({ size }) => {
           align="end"
           className="no-scrollbar max-h-[40vh] min-w-[14rem] overflow-scroll rounded-2xl"
         >
-          {/* Supported Chains */}
-          {/* {supportedChains.map((chain) => (
-            <DropdownMenuItem
-              disabled={chain.network === activeChain?.network}
-              className={
-                chain.network !== activeChain?.network ? "cursor-pointer" : ""
-              }
-              key={chain.network}
-              onClick={async () => {
-                await switchActiveChain?.(chain)
-                toast.success(`Switched to ${chain.name}`)
-              }}
-            >
-              <div className="flex w-full items-center justify-between gap-2">
-                <p>{chain.name}</p>
-                {chain.network === activeChain?.network && (
-                  <AiOutlineCheckCircle className="shrink-0" size={15} />
-                )}
-              </div>
-            </DropdownMenuItem>
-          ))} */}
-
           {/* Available Accounts/Wallets */}
           <DropdownMenuSeparator />
           {(accounts || []).map((acc) => {
@@ -215,9 +150,9 @@ export const ConnectButton: FC<ConnectButtonProps> = ({ size }) => {
                     <AccountName account={acc} />
                     <p className="text-xs">{truncatedEncodedAddress}</p>
                   </div>
-                  {acc.address === activeAccount?.address &&
-                    // <AiOutlineCheckCircle className="shrink-0" size={15} />
-                    "circle"}
+                  {acc.address === activeAccount?.address && (
+                    <CheckCircle size={16} />
+                  )}
                 </div>
               </DropdownMenuItem>
             )
@@ -229,10 +164,7 @@ export const ConnectButton: FC<ConnectButtonProps> = ({ size }) => {
             className="cursor-pointer"
             onClick={() => disconnect?.()}
           >
-            <div className="flex gap-2">
-              {/* <AiOutlineDisconnect size={18} /> */}
-              Disconnect
-            </div>
+            <div className="flex gap-2">Disconnect</div>
           </DropdownMenuItem>
         </DropdownMenuContent>
       </DropdownMenu>
@@ -245,18 +177,5 @@ export interface AccountNameProps {
 }
 
 export const AccountName: FC<AccountNameProps> = ({ account, ...rest }) => {
-  // const { activeChain } = useInkathon()
-  // const doResolveAddress = useMemo(
-  //   () =>
-  //     Object.values(SupportedChainId).includes(
-  //       activeChain?.network as SupportedChainId
-  //     ),
-  //   [activeChain?.network]
-  // )
-  // const { primaryDomain } = useResolveAddressToDomain(
-  //   doResolveAddress ? account?.address : undefined,
-  //   { chainId: activeChain?.network }
-  // )
-
   return <div>{account.name}</div>
 }
