+++ components/providers.tsx
@@ -2,36 +2,15 @@
 
 import { ChainProvider } from "@/providers/chain-provider"
 import { TooltipProvider } from "@radix-ui/react-tooltip"
-import {
-  SubstrateChain,
-  UseInkathonProvider,
-  development,
-  getSubstrateChain,
-  rococo,
-} from "@scio-labs/use-inkathon"
+import { UseInkathonProvider } from "@scio-labs/use-inkathon"
 import { QueryClient, QueryClientProvider } from "react-query"
+
+import { siteConfig } from "@/config/site"
 
 import { ThemeProvider } from "./theme-provider"
 
-const polkadotRelay: SubstrateChain = {
-  network: "Polkadot",
-  name: "Polkadot Relay Chain",
-  rpcUrls: ["wss://rpc.polkadot.io"],
-  ss58Prefix: 0,
-  testnet: false,
-}
-
-const rococoTestnet: SubstrateChain = {
-  network: "Rococo",
-  name: "Rococo Testnet",
-  rpcUrls: ["wss://rococo-rpc.polkadot.io"],
-  ss58Prefix: 42,
-  testnet: true,
-}
-
 export function Providers({ children }: { children: React.ReactNode }) {
   const queryClient = new QueryClient()
-  const defaultChain = rococoTestnet || development
 
   return (
     <ThemeProvider attribute="class" defaultTheme="system" enableSystem>
@@ -39,7 +18,7 @@ export function Providers({ children }: { children: React.ReactNode }) {
         <ChainProvider>
           <UseInkathonProvider
             appName="Polkadot Weigher"
-            defaultChain={defaultChain}
+            defaultChain={siteConfig.defaultChain}
             // connectOnInit={true}
           >
             <TooltipProvider>{children}</TooltipProvider>
