+++ config/site.ts
@@ -1,11 +1,29 @@
 import { DateRange } from "@/common/types"
+import { SubstrateChain, development } from "@scio-labs/use-inkathon"
 
 export type SiteConfig = typeof siteConfig
+
+const polkadotRelay: SubstrateChain = {
+  network: "Polkadot",
+  name: "Polkadot Relay Chain",
+  rpcUrls: ["wss://rpc.polkadot.io"],
+  ss58Prefix: 0,
+  testnet: false,
+}
+
+const rococoTestnet: SubstrateChain = {
+  network: "Rococo",
+  name: "Rococo Testnet",
+  rpcUrls: ["wss://rococo-rpc.polkadot.io"],
+  ss58Prefix: 42,
+  testnet: true,
+}
 
 export const siteConfig = {
   name: "Polkadot Weigher",
   description: "Displays utilization of Polkadot parachains.",
   backendUrl: "http://35.233.43.85:8000",
+  defaultChain: rococoTestnet || development,
   blockExplorer: "https://rococo.subscan.io/extrinsic/",
   subscriptionCost: "1000000000", // cost to register a parachain
   mainNav: [
